GuardScopeAnalyzer/requirements.txt
tabulate
argparse
boto3
tqdm
python-cfonts